requirements.txt
pyinstaller==3.6
pycryptodome==3.9.7
